AMPOPS-38 Merging and Branch Related Tasks AMP 1.16.2 -> AMP 1.16.6 merged into AMP 2.0
git-svn-id: file:///svn/branches/AMP_2_0_RELEASE@16041 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/repository/widget/view/indicatorchartwidgets.jsp
@@ -77,7 +77,7 @@
 								<digi:trn key="gis:editLink">Edit</digi:trn>
 							</a>
 							&nbsp;
-							<a href="/widget/indicatorchartwidgets.do~actType=delete~widgetId=${wi.id}">
+                                                        <a href="/widget/indicatorchartwidgets.do~actType=delete~widgetId=${wi.id}" onclick="return confirm('<digi:trn>Do you want to delete this widget?</digi:trn>')">
 								<img border="0" src='<digi:file src="images/deleteIcon.gif"/>'>
 							</a>
 						</td>
